% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_progress.R
\name{add_progress}
\alias{add_progress}
\title{Inject Progress Updates to an R Expression}
\usage{
add_progress(expr, substitute = TRUE)
}
\arguments{
\item{expr}{An \R expression.}

\item{substitute}{If TRUE, argument \code{expr} is \code{\link[base:substitute]{base::substitute()}}:ed.}
}
\description{
Inject Progress Updates to an R Expression
}
\examples{
with_progress(add_progress({
  x <- 1:10
  y <- slow_sum(x)
  z <- slow_sum(x^2)
  mu <- mean(z - y^2)
}))
print(mu)
}
