% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progression_handlers.R
\name{beepr_handler}
\alias{beepr_handler}
\title{Auditory Progression Feedback}
\usage{
beepr_handler(setup = 2L, update = 10L, done = 11L,
  intrusiveness = getOption("progressr.intrusiveness.auditory", 10), ...)
}
\description{
Auditory Progression Feedback
}
