% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler_brrr.R
\name{handler_brrr}
\alias{handler_brrr}
\title{Progression Handler: Progress Reported as 'BRRR' Sounds (Audio)}
\usage{
handler_brrr(
  initiate = "takeoff",
  update = "bigsean4",
  finish = "khaled3",
  interrupt = "busta",
  intrusiveness = getOption("progressr.intrusiveness.audio", 5),
  target = "audio",
  ...
)
}
\arguments{
\item{initiate, update, finish, interrupt}{(integer or character) Indices or
names of \code{skrrrahh()} sounds to play when progress starts, is updated,
completes, or is interrupted. For silence, use \code{NA_integer_}.}

\item{intrusiveness}{(numeric) A non-negative scalar on how intrusive
(disruptive) the reporter to the user.}

\item{target}{(character vector) Specifies where progression updates are
rendered.}

\item{\ldots}{Additional arguments passed to \code{\link[=make_progression_handler]{make_progression_handler()}}.}
}
\description{
A progression handler for \code{skrrrahh()} of the \pkg{BRRR} package.
}
\section{Requirements}{

This progression handler requires the \pkg{BRRR} package, which is only
available from \url{https://github.com/brooke-watson/BRRR}.  This can be
installed as \code{remotes::install_github("brooke-watson/BRRR")}.
}

\examples{
pkg <- "BRRR"
if (requireNamespace(pkg, quietly = TRUE)) {

  handlers("brrr")
  with_progress({ y <- slow_sum(1:10) })
  print(y)
  
}
}
