% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler_debug.R
\name{handler_debug}
\alias{handler_debug}
\title{Textual Progression Feedback for Debug Purposes}
\usage{
handler_debug(
  interval = getOption("progressr.interval", 0),
  intrusiveness = getOption("progressr.intrusiveness.debug", 0),
  target = "terminal",
  ...
)
}
\arguments{
\item{interval}{(numeric) The minimum time (in seconds) between
successive progression updates from this handler.}

\item{intrusiveness}{(numeric) A non-negative scalar on how intrusive
(disruptive) the reporter to the user.}

\item{target}{(character vector) Specifies where progression updates are
rendered.}

\item{\ldots}{Additional arguments passed to \code{\link[=make_progression_handler]{make_progression_handler()}}.}
}
\description{
Textual Progression Feedback for Debug Purposes
}
\examples{
handlers("debug")
with_progress({ y <- slow_sum(1:10) })
print(y)
}
