% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler_filesize.R
\name{handler_filesize}
\alias{handler_filesize}
\title{Progression Updates Reflected as the Size of a File}
\usage{
handler_filesize(
  file = "default.progress",
  intrusiveness = getOption("progressr.intrusiveness.file", 5),
  target = "file",
  ...
)
}
\arguments{
\item{file}{(character) A filename.}

\item{intrusiveness}{(numeric) A non-negative scalar on how intrusive
(disruptive) the reporter to the user.}

\item{target}{(character vector) Specifies where progression updates are
rendered.}

\item{\ldots}{Additional arguments passed to \code{\link[=make_progression_handler]{make_progression_handler()}}.}
}
\description{
Progression Updates Reflected as the Size of a File
}
\examples{
\donttest{\dontrun{
handlers(handler_filesize(file = "myscript.progress"))
with_progress(y <- slow_sum(1:100))
}}

}
