% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler_jgauge.R
\name{handler_jgauge}
\alias{handler_jgauge}
\title{Progression Handler: Progress Reported as "JGauge" Incremental Tones (Audio)}
\usage{
handler_jgauge(
  freq_range = c(220, 880),
  intrusiveness = getOption("progressr.intrusiveness.audio", 5),
  target = "audio",
  ...
)
}
\arguments{
\item{freq_range}{(numeric vector of length two) Lower and upper
frequencies (Hz) tones representing 0\% and 100\% progress, respectively.}

\item{intrusiveness}{(numeric) A non-negative scalar on how intrusive
(disruptive) the reporter to the user.}

\item{target}{(character vector) Specifies where progression updates are
rendered.}

\item{\ldots}{Additional arguments passed to \code{\link[=make_progression_handler]{make_progression_handler()}}.}
}
\description{
A progression handler that reports progress using incremental tones
as JGauge Progress Bar Monitor (1).
}
\section{Requirements}{

This progression handler requires an audio player.
}

\examples{
handlers("jgauge")
with_progress({ y <- slow_sum(1:10) })
print(y)

}
\references{
\enumerate{
\item Doug Lee, JGauge Progress Bar Monitoring Scripts For JAWS 2018 and
Later, \url{https://www.dlee.org/jgauge/}
}
}
