% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler_tutorial.R
\name{handler_tutorial}
\alias{handler_tutorial}
\title{A progressr handler useful for Futureverse tutorials}
\usage{
handler_tutorial(
  ...,
  spec = c("progress", "cli"),
  type = "time",
  clear = FALSE
)
}
\arguments{
\item{spec}{The progress handler to use.}

\item{type}{Controls which predefined handler format string to use.}

\item{clear}{(logical) If TRUE, any output, typically visual, produced
by a reporter will be cleared/removed upon completion, if possible.}

\item{\ldots}{Arguments passed to the selected progress handler.}
}
\value{
A function of class \code{progression_handler} that takes a
\link{progression} condition as its first and only argument.
}
\description{
A progressr handler useful for Futureverse tutorials
}
\details{
The default is to use the \pkg{progress} package. Contrary to the
\pkg{cli} package, the \pkg{progress} package has the capacity to
display \emph{all} progress updates, even when receiving progress updates
at a high rate from, for example, parallel processing.
}
\examples{
# Display progress and keep the progress output when done
handlers("tutorial")

# Display progress and keep every progress reported
handlers("tutorial", "newline")

}
