% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler_void.R
\name{handler_void}
\alias{handler_void}
\title{Progression Handler: No Progress Report}
\usage{
handler_void(intrusiveness = 0, target = "void", enable = FALSE, ...)
}
\arguments{
\item{intrusiveness}{(numeric) A non-negative scalar on how intrusive
(disruptive) the reporter to the user.}

\item{target}{(character vector) Specifies where progression updates are
rendered.}

\item{enable}{(logical) If FALSE, then progress is not reported.}

\item{\ldots}{Additional arguments passed to \code{\link[=make_progression_handler]{make_progression_handler()}}.}
}
\description{
Progression Handler: No Progress Report
}
\details{
This progression handler gives not output - it is invisible and silent.
}
\examples{
\donttest{\dontrun{
handlers(handler_void())
with_progress(y <- slow_sum(1:100))
print(y)
}}

}
