% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progression_handlers.R
\name{newline_handler}
\alias{newline_handler}
\title{Textual Progression Feedback that outputs a Newline}
\usage{
newline_handler(symbol = "\\n", file = stderr(),
  intrusiveness = getOption("progressr.intrusiveness.debug", 0), ...)
}
\arguments{
\item{symbol}{(character string) The character symbol to be outputted,
which by default is the ASCII NL character (\code{'\n'} = \code{'\013'}) character.}

\item{file}{(connection) A \link[base:connection]{base::connection} to where output should be sent.}

\item{intrusiveness}{(numeric) A non-negative scalar on how intrusive
(disruptive) the reporter to the user.}

\item{\ldots}{Additional arguments passed to \code{\link[=progression_handler]{progression_handler()}}.}
}
\description{
Textual Progression Feedback that outputs a Newline
}
