% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progression_handlers.R
\name{pbmcapply_handler}
\alias{pbmcapply_handler}
\title{Visual Progression Feedback}
\usage{
pbmcapply_handler(file = stderr(), style = "ETA", substyle = 3L,
  intrusiveness = getOption("progressr.intrusiveness.terminal", 10), ...)
}
\description{
Visual Progression Feedback
}
