% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress}
\alias{progress}
\title{Creates and Signals a Progression Condition}
\usage{
progress(..., call = sys.call())
}
\arguments{
\item{call}{(expression) A call expression.}

\item{\ldots}{Arguments pass to \code{\link[=progression]{progression()}}.}
}
\value{
A \link[base:conditions]{base::condition} of class \code{progression}.
}
\description{
\emph{WARNING:} \code{progress()} is deprecated - don't use.
}
\seealso{
To create a progression condition, use \code{\link[=progression]{progression()}}.
To signal a progression condition, use \code{\link[base:conditions]{base::signalCondition()}}.
}
\keyword{internal}
