% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_aggregator.R
\name{progress_aggregator}
\alias{progress_aggregator}
\title{Aggregate Progression Conditions}
\usage{
progress_aggregator(progress)
}
\arguments{
\item{progress}{A \link{progressor} function.}
}
\value{
A function of class \code{progress_aggregator}.
}
\description{
Aggregate Progression Conditions
}
\examples{
library(progressr)

message("progress_aggregator() ...")

with_progress({
  progress <- progressor(steps = 4L)
  relay_progress <- progress_aggregator(progress)
  progress()
  relay_progress(slow_sum(1:3))
  relay_progress(slow_sum(1:10))
  progress()
})

message("progress_aggregator() ... done")
}
\keyword{internal}
