% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plyr_progress_progressr.R
\name{progress_progressr}
\alias{progress_progressr}
\title{Use Progressr with Plyr Map-Reduce Functions}
\usage{
progress_progressr(...)
}
\arguments{
\item{\ldots}{Not used.}
}
\value{
A named \link[base:list]{base::list} that can be passed as argument \code{.progress}
to any of \pkg{plyr} function accepting that argument.
}
\description{
A "progress bar" for \pkg{plyr}'s \code{.progress} argument.
}
\examples{
with_progress({
  y <- plyr::l_ply(1:10, function(...) Sys.sleep(0.1), .progress = "progressr")
})

}
