% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progression.R
\name{progression}
\alias{progression}
\title{A Progression Condition}
\usage{
progression(message = character(0L), amount = 1, step = NULL,
  time = Sys.time(), ..., type = c("update", "setup", "done"),
  class = NULL, call = NULL)
}
\arguments{
\item{message}{(character) A progress message.}

\item{amount}{(numeric) The total amount of progress made.}

\item{step}{(character or integer) The step completed.}

\item{time}{(POSIXct) A timestamp.}

\item{type}{Type of progression made.}

\item{class}{(character) Zero or more class names to prepend.}

\item{call}{(expression) A call expression.}

\item{\ldots}{Additional named elements.}
}
\value{
A \link[base:condition]{base::condition} of class \code{progression}.
}
\description{
A Progression Condition
}
\seealso{
To signal a progression condition, use \code{\link[base:signalCondition]{base::signalCondition()}}.
To create and signal a progression condition at once, use \code{\link[=progress]{progress()}}.
}
