% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progression_handler.R
\name{progression_handler}
\alias{progression_handler}
\title{Creates a Progression Handler}
\usage{
progression_handler(name, reporter = environment(), handler = NULL,
  times = getOption("progressr.times", +Inf),
  interval = getOption("progressr.interval", 0))
}
\arguments{
\item{name}{(character) Name of progression handler.}

\item{handler}{(environment) A reporter environment.}

\item{times}{(integer) The maximum number of times this handler
should report progression updates.}

\item{interval}{(numeric) The minimum time (in seconds) between
successive progression updates from this handler.}

\item{handler}{(function) Function take a \code{progression} condition
as the first argument.}
}
\value{
A function of class \code{progression_handler}.
}
\description{
Creates a Progression Handler
}
