% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slow_sqrt.R
\name{slow_sqrt}
\alias{slow_sqrt}
\title{Slowly Calculate Square Roots of Elements}
\usage{
slow_sqrt(
  x,
  delay = getOption("progressr.demo.delay", 1),
  stdout = getOption("progressr.demo.stdout", FALSE),
  message = getOption("progressr.demo.message", TRUE),
  sticky = getOption("progressr.demo.sticky", TRUE)
)
}
\arguments{
\item{x}{Numeric vector to calculate square roots for.}

\item{delay}{Delay in seconds after each calculation.}

\item{stdout}{If TRUE, then a text is outputted to the standard output
per element.}

\item{message}{If TRUE, then a message is outputted per element.}

\item{sticky}{If TRUE, then a "sticky" message is outputted every
ten element.}
}
\value{
The square roots of all elements in \code{x} as a numeric vector.
}
\description{
Slowly Calculate Square Roots of Elements
}
\section{Progress updates}{

This function signals \link{progression} conditions as it progresses.
}

\keyword{internal}
