% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slow_sum.R
\name{slow_sum}
\alias{slow_sum}
\title{Slowly Calculate Sum of Elements}
\usage{
slow_sum(x, delay = getOption("delay", 0.05))
}
\arguments{
\item{x}{Numeric vector to sum}

\item{delay}{Delay in seconds after each addition.}
}
\value{
The sum of all elements in \code{x}.
}
\description{
Slowly Calculate Sum of Elements
}
\section{Progress updates}{

This function signals \link{progression} conditions as it progresses.
}

