% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slow_sum_p.R
\name{slow_sum_p}
\alias{slow_sum_p}
\title{Slowly Calculate Sum of Elements}
\usage{
slow_sum_p(
  x,
  delay = getOption("progressr.demo.delay", 1),
  stdout = getOption("progressr.demo.stdout", FALSE),
  message = getOption("progressr.demo.message", TRUE),
  sticky = getOption("progressr.demo.sticky", TRUE),
  progress = getOption("progressr.demo.progress", TRUE)
)
}
\arguments{
\item{x}{Numeric vector to sum}

\item{delay}{Delay in seconds after each addition.}

\item{stdout}{If TRUE, then a text is outputted to the standard output
per element.}

\item{message}{If TRUE, then a message is outputted per element.}

\item{sticky}{If TRUE, then a "sticky" message is outputted every
ten element.}

\item{progress}{If TRUE, then a progressor is created, otherwise a void
progressor that does nothing is used.}
}
\value{
The sum of all elements in \code{x}.
}
\description{
Slowly Calculate Sum of Elements
}
\section{Progress updates}{

This function signals \link{progression} conditions as it progresses.
}

\keyword{internal}
