% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progression_handlers.R
\name{txtprogressbar_handler}
\alias{txtprogressbar_handler}
\title{Visual Progression Feedback}
\usage{
txtprogressbar_handler(file = stderr(), style = 3L,
  intrusiveness = getOption("progressr.intrusiveness.terminal", 1), ...)
}
\description{
Visual Progression Feedback
}
