% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_progress.R
\name{with_progress}
\alias{with_progress}
\title{Evaluate Expression while Handling Progress Updates}
\usage{
with_progress(expr, handler = getOption("progressr.handler",
  txtprogressbar_handler()))
}
\arguments{
\item{expr}{An \R expression to evaluate.}

\item{handler}{(function) A progression handler.
If @NULL, progress updates are ignored.}
}
\description{
Evaluate Expression while Handling Progress Updates
}
