% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progression_handlers.R
\name{ascii_alert_handler}
\alias{ascii_alert_handler}
\title{Auditory Progression Feedback}
\usage{
ascii_alert_handler(symbol = "\\a", file = stderr(),
  intrusiveness = getOption("progressr.intrusiveness.auditory", 10), ...)
}
\description{
Auditory Progression Feedback
}
