% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progression_handlers.R
\name{ascii_alert_handler}
\alias{ascii_alert_handler}
\title{Auditory Progression Feedback}
\usage{
ascii_alert_handler(symbol = "\\a", file = stderr(),
  intrusiveness = getOption("progressr.intrusiveness.auditory", 10), ...)
}
\arguments{
\item{symbol}{(character string) The character symbol to be outputted,
which by default is the ASCII BEL character (\code{'\a'} = \code{'\007'}) character.}

\item{file}{(connection) A \link[base:connection]{base::connection} to where output should be sent.}

\item{intrusiveness}{(numeric) A non-negative scalar on how intrusive
(disruptive) the reporter to the user.}

\item{\ldots}{Additional arguments passed to \code{\link[=progression_handler]{progression_handler()}}.}
}
\description{
A progression handler based on \code{cat("\a", file=stderr())}.
}
\examples{
if (requireNamespace("utils", quietly = TRUE)) local({
  oopts <- options(progressr.clear = FALSE,
                   progressr.handlers = ascii_alert_handler)
  on.exit(oopts)

  x <- 1:10
  with_progress({ y <- slow_sum(x) })
  
  print(y)
})
}
