% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progression_handlers.R
\name{ascii_alert_handler}
\alias{ascii_alert_handler}
\title{Auditory Progression Feedback}
\usage{
ascii_alert_handler(symbol = "\\a", ...,
  times = getOption("progressr.times", +Inf),
  interval = getOption("progressr.interval", 0.5), file = stderr(),
  enable = interactive())
}
\description{
Auditory Progression Feedback
}
