% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progression_handlers.R
\name{beepr_handler}
\alias{beepr_handler}
\title{Auditory Progression Feedback}
\usage{
beepr_handler(initiate = 2L, update = 10L, finish = 11L,
  intrusiveness = getOption("progressr.intrusiveness.auditory", 10), ...)
}
\arguments{
\item{initiate, update, finish}{(integer) Indices of \code{\link[beepr:beep]{beepr::beep()}} sounds to
play when progress starts, is updated, and completes.  For silence, use \code{NA_integer_}.}

\item{intrusiveness}{(numeric) A non-negative scalar on how intrusive
(disruptive) the reporter to the user.}

\item{\ldots}{Additional arguments passed to \code{\link[=progression_handler]{progression_handler()}}.}
}
\description{
A progression handler for \code{\link[beepr:beep]{beepr::beep()}}.
}
\examples{
if (requireNamespace("utils", quietly = TRUE)) local({
  oopts <- options(progressr.clear = FALSE,
                   progressr.handlers = beepr_handler)
  on.exit(oopts)

  x <- 1:10
  with_progress({ y <- slow_sum(x) })
  
  print(y)
})
}
