% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progression_handlers.R
\name{beepr_handler}
\alias{beepr_handler}
\title{Auditory Progression Feedback}
\usage{
beepr_handler(setup = 2L, update = 10L, done = 11L,
  times = getOption("progressr.times", +Inf),
  interval = getOption("progressr.interval", 0.5),
  enable = interactive(), ...)
}
\description{
Auditory Progression Feedback
}
