% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progression_handlers.R
\name{beepr_handler}
\alias{beepr_handler}
\title{Auditory Progression Feedback}
\usage{
beepr_handler(setup_sound = 2L, update_sound = 10L, done_sound = 11L,
  enable = interactive(), times = getOption("progressr.times", +Inf),
  interval = getOption("progressr.interval", 0.5), ...)
}
\description{
Auditory Progression Feedback
}
