% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progression_handlers.R
\name{filesize_handler}
\alias{filesize_handler}
\title{Progression Updates Reflected as the Size of a File}
\usage{
filesize_handler(file = "default.progress",
  intrusiveness = getOption("progressr.intrusiveness.file", 5), ...)
}
\arguments{
\item{file}{(character) A filename.}

\item{intrusiveness}{(numeric) A non-negative scalar on how intrusive
(disruptive) the reporter to the user.}

\item{\ldots}{Additional arguments passed to \code{\link[=progression_handler]{progression_handler()}}.}
}
\description{
Progression Updates Reflected as the Size of a File
}
\examples{
\donttest{\dontrun{
options(progressr.handlers = filesize_handler(file = "myscript.progress"))
with_progress(y <- slow_sum(1:100))
}}

}
