% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_progression_handler.R
\name{global_progression_handler}
\alias{global_progression_handler}
\title{A Global Calling Handler For 'progression':s}
\usage{
global_progression_handler(condition)
}
\arguments{
\item{progression}{A \link{progression} conditions.}
}
\value{
Nothing.
}
\description{
A Global Calling Handler For 'progression':s
}
\section{Requirements}{

This function requires R (>= 4.0.0) - the version in which global calling
handlers where introduces.
}

\keyword{internal}
