% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler_cli.R
\name{handler_cli}
\alias{handler_cli}
\title{Progression Handler: Progress Reported via 'cli' Progress Bars (Text) in the Terminal}
\usage{
handler_cli(
  show_after = 0,
  intrusiveness = getOption("progressr.intrusiveness.terminal", 1),
  target = "terminal",
  ...
)
}
\arguments{
\item{show_after}{(numeric) Number of seconds to wait before displaying
the progress bar.}

\item{intrusiveness}{(numeric) A non-negative scalar on how intrusive
(disruptive) the reporter to the user.}

\item{target}{(character vector) Specifies where progression updates are
rendered.}

\item{\ldots}{Additional arguments passed to \code{\link[=make_progression_handler]{make_progression_handler()}}.}
}
\description{
A progression handler for \code{\link[cli:cli_progress_bar]{cli::cli_progress_bar()}}.
}
\section{Requirements}{

This progression handler requires the \pkg{cli} package.
}

\section{Appearance}{

Below is how this progress handler renders by default at 0\%, 30\% and 99\%
progress:

With \code{handlers(handler_cli())}:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{                                 0\% | ETA: ?
xxxxxxxxxxxxxxxxx               60\% | ETA: 19s
xxxxxxxxxxxxxxxxxxxxxxxxxxxxx   99\% | ETA:  0s
}\if{html}{\out{</div>}}
}

\examples{
if (requireNamespace("cli", quietly = TRUE)) local({
  ## with_progress() on a "cli" handler gives an error if a
  ## global progressing handler is set.
  if (getRversion() >= "4.0.0" && handlers(global = NA)) {
    handlers(global = FALSE)
    on.exit(handlers(global = TRUE))
  }
  
  handlers(handler_cli(format = "{cli::pb_spin} {cli::pb_bar} {cli::pb_percent} {cli::pb_status}"))
  with_progress({ y <- slow_sum(1:10) })
  print(y)
})
}
