% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler_pbcol.R
\name{handler_pbcol}
\alias{handler_pbcol}
\title{Progression Handler: Progress Reported as an ANSI Background Color in the Terminal}
\usage{
handler_pbcol(
  adjust = 0,
  pad = 1L,
  text_col = "white",
  done_col = "blue",
  todo_col = "cyan",
  intrusiveness = getOption("progressr.intrusiveness.terminal", 1),
  target = "terminal",
  ...
)
}
\arguments{
\item{adjust}{(numeric) The adjustment of the progress update,
where \code{adjust = 0} positions the message to the very left, and
\code{adjust = 1} positions the message to the very right.}

\item{pad}{(integer) Amount of padding on each side of the message,
where padding is done by spaces.}

\item{text_col, done_col, todo_col}{(character string) The \pkg{crayon}
foreground and background colors used for the progress bar, where
\code{text_col} is used for all of the progress bar, \code{done_col} is used for the
part of the progress bar that is already done and \code{todo_col} for what
remains.}

\item{intrusiveness}{(numeric) A non-negative scalar on how intrusive
(disruptive) the reporter to the user.}

\item{target}{(character vector) Specifies where progression updates are
rendered.}

\item{\ldots}{Additional arguments passed to \code{\link[=make_progression_handler]{make_progression_handler()}}.}
}
\description{
Progression Handler: Progress Reported as an ANSI Background Color in the Terminal
}
\section{Requirements}{

This progression handler requires the \pkg{crayon} package.
}

\examples{
handlers(handler_pbcol)
with_progress({ y <- slow_sum(1:10) })
print(y)
}
