% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler_winprogressbar.R
\name{handler_winprogressbar}
\alias{handler_winprogressbar}
\title{Progression Handler: Progress Reported as a MS Windows Progress Bars in the GUI}
\usage{
handler_winprogressbar(
  intrusiveness = getOption("progressr.intrusiveness.gui", 1),
  target = "gui",
  ...
)
}
\arguments{
\item{intrusiveness}{(numeric) A non-negative scalar on how intrusive
(disruptive) the reporter to the user.}

\item{target}{(character vector) Specifies where progression updates are
rendered.}

\item{\ldots}{Additional arguments passed to \code{\link[=make_progression_handler]{make_progression_handler()}}.}
}
\description{
A progression handler for \code{winProgressBar()} in the \pkg{utils} package.
}
\section{Requirements}{

This progression handler requires MS Windows.
}

