% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{handlers}
\alias{handlers}
\title{Get and Set the List of Progression Handlers}
\usage{
handlers(..., on_missing = c("error", "warning", "ignore"))
}
\arguments{
\item{\dots}{One or more progression handlers.  Alternatively, this
functions accepts also a single list of progression handlers as input.
If this list is empty, then an empty set of progression handlers will
be set.}

\item{on_missing}{(character) If \code{"error"}, an error is thrown if one of
the progression handlers does not exists.  If \code{"warning"}, a warning
is produces and the missing handlers is ignored.  If \code{"ignore"}, the
missing handlers is ignored.}
}
\value{
(invisibly) the previous list of progression handlers set.
If no arguments are specified, then the current set of progression
handlers is returned.
}
\description{
Get and Set the List of Progression Handlers
}
\details{
This function provides a convenient alternative for getting and setting
option \option{progressr.handlers}.
}
\examples{
hdlrs <- "txtprogressbar"
if (requireNamespace("beepr", quietly = TRUE)) hdlrs <- c(hdlrs, "beepr")
handlers(hdlrs)

with_progress({ y <- slow_sum(10) })
print(y)
}
