% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progression_handlers.R
\name{notifier_handler}
\alias{notifier_handler}
\title{Operating-System Specific Progression Feedback}
\usage{
notifier_handler(setup = 2L, update = 10L, done = 11L,
  enable = interactive(), times = getOption("progressr.times", +Inf),
  interval = getOption("progressr.interval", 5), ...)
}
\description{
Operating-System Specific Progression Feedback
}
