% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progression_handlers.R
\name{notifier_handler}
\alias{notifier_handler}
\title{Operating-System Specific Progression Feedback}
\usage{

  notifier_handler(intrusiveness = getOption("progressr.intrusiveness.notifier",
  10), ...)
}
\arguments{
\item{intrusiveness}{(numeric) A non-negative scalar on how intrusive
(disruptive) the reporter to the user.}

\item{\ldots}{Additional arguments passed to \code{\link[=progression_handler]{progression_handler()}}.}
}
\description{
A progression handler for \code{\link[notifier:notify]{notifier::notify()}}.
}
\examples{
if (requireNamespace("notifier", quietly = TRUE)) local({
  oopts <- options(progressr.clear = FALSE,
                   progressr.handlers = notifier_handler)
  on.exit(oopts)

  x <- 1:10
  with_progress({ y <- slow_sum(x) })
  
  print(y)
})
}
