% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progression_handlers.R
\name{notifier_handler}
\alias{notifier_handler}
\title{Operating-System Specific Progression Feedback}
\usage{

  notifier_handler(intrusiveness = getOption("progressr.intrusiveness.notifier",
  10), target = "gui", ...)
}
\arguments{
\item{intrusiveness}{(numeric) A non-negative scalar on how intrusive
(disruptive) the reporter to the user.}

\item{target}{(character vector) Specifies where progression updates are
rendered.}

\item{\ldots}{Additional arguments passed to \code{\link[=make_progression_handler]{make_progression_handler()}}.}
}
\description{
A progression handler for \code{\link[notifier:notify]{notifier::notify()}}.
}
\examples{
if (requireNamespace("notifier", quietly = TRUE)) {
  handlers("notifier")
  with_progress({ y <- slow_sum(1:10) })
  print(y)
}
}
