% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress}
\alias{progress}
\title{Creates and Signals a Progression Condition}
\usage{
progress(..., call = sys.call())
}
\arguments{
\item{call}{(expression) A call expression.}

\item{\ldots}{Arguments pass to \code{\link[=progression]{progression()}}.}
}
\value{
A \link[base:conditions]{base::condition} of class \code{progression}.
}
\description{
Creates and Signals a Progression Condition
}
\seealso{
To signal a progression condition, use \code{\link[base:conditions]{base::signalCondition()}}.
To create and signal a progression condition at once, use \code{\link[=progress]{progress()}}.
}
\keyword{internal}
