% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progression_handlers.R
\name{progress_handler}
\alias{progress_handler}
\title{Visual Progression Feedback}
\usage{
progress_handler(show_after = 0,
  intrusiveness = getOption("progressr.intrusiveness.terminal", 10), ...)
}
\description{
Visual Progression Feedback
}
