% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progression_handlers.R
\name{progress_handler}
\alias{progress_handler}
\title{Visual Progression Feedback}
\usage{
progress_handler(show_after = 0,
  intrusiveness = getOption("progressr.intrusiveness.terminal", 1), ...)
}
\arguments{
\item{show_after}{(numeric) Number of seconds to wait before displaying the progress bar.}

\item{intrusiveness}{(numeric) A positive scalar on how intrusive
(disruptive) the reporter to the user.}

\item{\ldots}{Additional arguments passed to \code{\link[=progression_handler]{progression_handler()}}.}
}
\description{
A progression handler for \code{\link[progress:progress_bar]{progress::progress_bar()}}.
}
