% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progression_handlers.R
\name{progress_handler}
\alias{progress_handler}
\title{Visual Progression Feedback}
\usage{
progress_handler(..., clear = FALSE, show_after = 0,
  times = getOption("progressr.times", +Inf),
  interval = getOption("progressr.interval", 0))
}
\description{
Visual Progression Feedback
}
