% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progression_handlers.R
\name{progress_handler}
\alias{progress_handler}
\title{Visual Progression Feedback}
\usage{
progress_handler(format = "[:bar] :percent :message", show_after = 0,
  intrusiveness = getOption("progressr.intrusiveness.terminal", 1), ...)
}
\arguments{
\item{format}{(character string) The format of the progress bar.}

\item{show_after}{(numeric) Number of seconds to wait before displaying
the progress bar.}

\item{intrusiveness}{(numeric) A non-negative scalar on how intrusive
(disruptive) the reporter to the user.}

\item{\ldots}{Additional arguments passed to \code{\link[=make_progression_handler]{make_progression_handler()}}.}
}
\description{
A progression handler for \code{\link[progress:progress_bar]{progress::progress_bar()}}.
}
\examples{
if (requireNamespace("progress", quietly = TRUE)) {
  handlers("progress")
  with_progress({ y <- slow_sum(1:10) })
  print(y)
}
}
