% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressor_plyr.R
\name{progress_progressr}
\alias{progress_progressr}
\title{Progressr Progression Updates in Plyr}
\usage{
progress_progressr(...)
}
\arguments{
\item{\ldots}{Not used.}
}
\value{
A named \link[base:list]{base::list} that can be passed as argument \code{.progress}
to any of \pkg{plyr} function accepting that argument.
}
\description{
A \pkg{plyr} progress "bar" that signals \link{progression} conditions.
}
\examples{
with_progress({
  y <- plyr::l_ply(1:10, function(...) Sys.sleep(0.1), .progress = "progressr")
})

}
