% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progression.R
\name{progression}
\alias{progression}
\title{A Progression Condition}
\usage{
progression(amount = 1, message = character(0L), step = NULL,
  time = Sys.time(), ..., type = c("update", "initiate", "finish"),
  class = NULL, progressor_uuid = NULL, progression_index = NULL,
  session_uuid = NULL, call = NULL)
}
\arguments{
\item{amount}{(numeric) The total amount of progress made.}

\item{message}{(character) A progress message.}

\item{step}{(character or integer) The step completed.}

\item{time}{(POSIXct) A timestamp.}

\item{type}{Type of progression made.}

\item{class}{(character) Zero or more class names to prepend.}

\item{progressor_uuid}{(character string) A character string that is unique
for the current progressor and the current \R session.}

\item{progression_index}{(integer) A non-negative integer that is}

\item{session_uuid}{(character string) A character string that is unique
for the current \R session.}

\item{call}{(expression) A call expression.}

\item{\ldots}{Additional named elements.}
}
\value{
A \link[base:condition]{base::condition} of class \code{progression}.
}
\description{
A Progression Condition
}
\seealso{
To signal a progression condition, use \code{\link[base:signalCondition]{base::signalCondition()}}.
To create and signal a progression condition at once, use \code{\link[=progress]{progress()}}.
}
