% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progression_handler.R
\name{progression_handler}
\alias{progression_handler}
\title{Creates a Progression Handler}
\usage{
progression_handler(name, reporter = list(), handler = NULL,
  enable = interactive(), times = getOption("progressr.times", +Inf),
  interval = getOption("progressr.interval", 0), intrusiveness = 1,
  clear = TRUE)
}
\arguments{
\item{name}{(character) Name of progression handler.}

\item{reporter}{(environment) A reporter environment.}

\item{handler}{(function) Function take a \code{progression} condition
as the first argument.}

\item{enable}{(logical) If FALSE, then progress is not reported.}

\item{times}{(integer) The maximum number of times this handler
should report progression updates.
If zero, then progress is not reported.}

\item{interval}{(numeric) The minimum time (in seconds) between
successive progression updates from this handler.}

\item{intrusiveness}{(numeric) A positive scalar on how intrusive
(disruptive) the reporter to the user.}

\item{clear}{(logical) If TRUE, any output, typically visual, produced
by a reporter will be cleared/removed upon completion, if possible.}
}
\value{
A function of class \code{progression_handler}.
}
\description{
Creates a Progression Handler
}
