% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressor.R
\name{progressor}
\alias{progressor}
\title{Create a Progressor Function}
\usage{
progressor(steps = NULL, along = NULL, label = NA_character_,
  initiate = TRUE, auto_finish = TRUE)
}
\arguments{
\item{steps}{(integer) Maximum number of steps.}

\item{along}{(vector; alternative) Corresponds to \code{steps = seq_along(along)}.}

\item{label}{(character) A label.}

\item{initiate}{(logical) If TRUE, the progressor will signal a
\link{progression} 'initiate' condition.}

\item{auto_finish}{(logical) If TRUE, then the progressor will signal a
\link{progression} 'finish' condition as soon as the last step has been reached.}
}
\value{
A function of class \code{progressor}.
}
\description{
Create a Progressor Function
}
