% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressor.R
\name{progressor}
\alias{progressor}
\title{Create a Progressor Function}
\usage{
progressor(steps, label = NA_character_, setup = TRUE,
  auto_done = TRUE)
}
\arguments{
\item{steps}{(integer) Maximum number of steps.}

\item{label}{(character) A label.}

\item{setup}{(logical) If TRUE, the progressor will signal a
\link{progression} 'setup' condition.}

\item{auto_done}{(logical) If TRUE, then the progressor will signal a
\link{progression} 'done' condition as soon as the last step has been reached.}
}
\value{
A function of class \code{progressor}.
}
\description{
Create a Progressor Function
}
