% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progression_handlers.R
\name{shiny_handler}
\alias{shiny_handler}
\title{Visual Progression Feedback}
\usage{
shiny_handler(
  intrusiveness = getOption("progressr.intrusiveness.gui", 1),
  target = "gui",
  ...
)
}
\arguments{
\item{intrusiveness}{(numeric) A non-negative scalar on how intrusive
(disruptive) the reporter to the user.}

\item{target}{(character vector) Specifies where progression updates are
rendered.}

\item{\ldots}{Additional arguments passed to \code{\link[=make_progression_handler]{make_progression_handler()}}.}
}
\description{
A progression handler for \pkg{shiny} and \code{\link[shiny:withProgress]{shiny::withProgress()}}.
}
\examples{
\donttest{\dontrun{
options(progressr.handlers = shiny_handler())
with_progress(y <- slow_sum(1:100))
}}

}
