% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slow_sum_p.R
\name{slow_sum_p}
\alias{slow_sum_p}
\alias{slow_sum}
\alias{slow_sqrt_p}
\alias{slow_sqrt}
\title{Slowly Calculate Sum or Square Roots of Elements}
\usage{
slow_sum_p(
  x,
  delay = getOption("progressr.demo.delay", 1),
  stdout = getOption("progressr.demo.stdout", FALSE),
  message = getOption("progressr.demo.message", TRUE),
  sticky = getOption("progressr.demo.sticky", TRUE),
  progress = getOption("progressr.demo.progress", TRUE)
)

slow_sum(
  x,
  delay = getOption("progressr.demo.delay", 1),
  stdout = FALSE,
  message = FALSE,
  sticky = FALSE,
  progress = FALSE
)

slow_sqrt_p(
  x,
  delay = getOption("progressr.demo.delay", 0.5),
  stdout = getOption("progressr.demo.stdout", FALSE),
  message = getOption("progressr.demo.message", TRUE),
  sticky = getOption("progressr.demo.sticky", TRUE),
  progress = getOption("progressr.demo.progress", TRUE)
)

slow_sqrt(
  x,
  delay = getOption("progressr.demo.delay", 0.5),
  stdout = FALSE,
  message = FALSE,
  sticky = FALSE,
  progress = FALSE
)
}
\arguments{
\item{x}{Numeric vector to sum or calculate square roots for.}

\item{delay}{Delay in seconds after each addition/calculation.}

\item{stdout}{If TRUE, then a text is outputted to the standard output
per element.}

\item{message}{If TRUE, then a message is outputted per element.}

\item{sticky}{If TRUE, then a "sticky" message is outputted every
ten element.}

\item{progress}{If TRUE, then a progressor is created, otherwise a void
progressor that does nothing is used.}
}
\value{
The sum or square roots of all elements in \code{x}.
}
\description{
Slowly Calculate Sum or Square Roots of Elements
}
\section{Progress updates}{

These functions signal \link{progression} conditions as they progress.
}

\keyword{internal}
