% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progression_handlers.R
\name{tkprogressbar_handler}
\alias{tkprogressbar_handler}
\title{Visual Progression Feedback}
\usage{

  tkprogressbar_handler(intrusiveness = getOption("progressr.intrusiveness.gui",
  10), ...)
}
\description{
Visual Progression Feedback
}
