% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progression_handlers.R
\name{tkprogressbar_handler}
\alias{tkprogressbar_handler}
\title{Visual Progression Feedback}
\usage{

  tkprogressbar_handler(intrusiveness = getOption("progressr.intrusiveness.gui",
  1), ...)
}
\arguments{
\item{intrusiveness}{(numeric) A non-negative scalar on how intrusive
(disruptive) the reporter to the user.}

\item{\ldots}{Additional arguments passed to \code{\link[=progression_handler]{progression_handler()}}.}
}
\description{
A progression handler for \code{\link[tcltk:tkProgressBar]{tcltk::tkProgressBar()}}.
}
