% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progression_handlers.R
\name{txtprogressbar_handler}
\alias{txtprogressbar_handler}
\title{Visual Progression Feedback}
\usage{
txtprogressbar_handler(..., times = getOption("progressr.times", +Inf),
  interval = getOption("progressr.interval", 0), file = stderr())
}
\description{
Visual Progression Feedback
}
