% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progression_handlers.R
\name{txtprogressbar_handler}
\alias{txtprogressbar_handler}
\title{Visual Progression Feedback}
\usage{
txtprogressbar_handler(style = 3L, file = stderr(),
  intrusiveness = getOption("progressr.intrusiveness.terminal", 1), ...)
}
\arguments{
\item{style}{(integer) The progress-bar style according to \code{\link[utils:txtProgressBar]{utils::txtProgressBar()}}.}

\item{file}{(connection) A \link[base:connection]{base::connection} to where output should be sent.}

\item{intrusiveness}{(numeric) A non-negative scalar on how intrusive
(disruptive) the reporter to the user.}

\item{\ldots}{Additional arguments passed to \code{\link[=progression_handler]{progression_handler()}}.}
}
\description{
A progression handler for \code{\link[utils:txtProgressBar]{utils::txtProgressBar()}}.
}
\examples{
if (requireNamespace("utils", quietly = TRUE)) local({
  oopts <- options(progressr.clear = FALSE,
                   progressr.handlers = txtprogressbar_handler)
  on.exit(oopts)

  x <- 1:10
  with_progress({ y <- slow_sum(x) })
  
  print(y)
})
}
