% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progression_handlers.R
\name{winprogressbar_handler}
\alias{winprogressbar_handler}
\title{Visual Progression Feedback for MS Windows}
\usage{

  winprogressbar_handler(intrusiveness = getOption("progressr.intrusiveness.gui",
  1), target = "gui", ...)
}
\arguments{
\item{intrusiveness}{(numeric) A non-negative scalar on how intrusive
(disruptive) the reporter to the user.}

\item{target}{(character vector) Specifies where progression updates are
rendered.}

\item{\ldots}{Additional arguments passed to \code{\link[=make_progression_handler]{make_progression_handler()}}.}
}
\description{
A progression handler for \code{\link[utils:winProgressBar]{utils::winProgressBar()}}.
}
