% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_progress.R
\name{with_progress}
\alias{with_progress}
\title{Evaluate Expression while Handling Progress Updates}
\usage{
with_progress(expr, renderer = renderer_txtprogressbar())
}
\arguments{
\item{expr}{An \R expression to evaluate.}

\item{renderer}{(function) A progression renderer.
If @NULL, progress updates are ignored.}
}
\description{
Evaluate Expression while Handling Progress Updates
}
