% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_progress.R
\name{with_progress}
\alias{with_progress}
\title{Evaluate Expression while Handling Progress Updates}
\usage{
with_progress(expr, handler = getOption("progressr.handler",
  txtprogressbar_handler()))
}
\arguments{
\item{expr}{An \R expression to evaluate.}

\item{handler}{(function) A progression handler.
If @NULL, progress updates are ignored.}
}
\description{
Evaluate Expression while Handling Progress Updates
}
\examples{
x <- 1:10

## Without progress updates
y <- slow_sum(x)


## Progress reported via txtProgressBar (default)
if (requireNamespace("utils", quietly = TRUE)) {
  with_progress({
    y <- slow_sum(x)
  })
}

## Progress reported via tcltk::tkProgressBar
if (requireNamespace("tcltk", quietly = TRUE)) {
  with_progress({
    y <- slow_sum(x)
  }, tkprogressbar_handler)
}

## Progress reported via progress::progress_bar)
if (requireNamespace("progress", quietly = TRUE)) {
  with_progress({
    y <- slow_sum(x)
  }, progress_handler)
}

## Progress reported via txtProgressBar + beepr::beep
if (requireNamespace("utils", quietly = TRUE) && requireNamespace("beepr", quietly = TRUE)) {
  with_progress({
    with_progress({
      y <- slow_sum(x)
    }, txtprogressbar_handler(style = 3L))
  }, beepr_handler)
}
}
